Different trials to see behavior of CSF.
The constant is set to obtain the correct pressure, but as soon as refinement is changed, pressure is changed (factor 4). The problem comes from kappa and grad phi....

Different grad phi has been tested: the unprolongated one leads to more instability (longer vibration of circle) whereas the prolongated one is more stable. Changin phixy by phiyx doesnt change the result.

Post processing has been tested with different epsilon and different sharpness, but it just re-set the circle, nothing happens.

Viscosity mu has no influence on pressure.

The smoothness of initial circle plays an important role and delivers kappa close to 50, since the interface is much smoother...but delta p gets lower.

Multiplying the constant by a factor X will give delta p times X.

Changing delta t delivers same result, but with more or less computational time...too low Dt increases instability and needs more time to solve.

The big issue, along with the false dp, is that the results are not the same when the mesh is refined. One finer mesh increases Dp with a factor 4.

It all boils down to the product kappa * Grad Phi in the right hand side. They depend on the mesh and on the interface Phi.

Next step: investigate more on this product kappa * Grad Phi...and try to understand why it works with BrackBill