FINALLY ENDED UP WITH THE CORRECT PRESSURE FOR THE EQUILIBRIUM ROD!!!!
The problem came from the rho factor in front of the B matrices!!!!!!! Now it is removed.

Consider also that the values of Paraview and the p max and p min values are far from being the relevant one to compare with papers!!

One has to go through the solution vector and take values from there to compare.

It also appeared that we should obtain the same solution independently from epsilon. Just have to wait a long time until the peaks are smeared out and the stationary solution is there.

Next step: calculate exactly the pressure drop and compare result