Here is a successful test of a falling drop.
Example from equilibrium drop reworked with realistic fluid properties (+corrected VOF class to enable variable viscosity in example 43, it was not the case, all previous commits were using mu=1 in example 43). Putting smago viscosity to 0.2 enables a reasonable result. Problem = can't compare result with literature, just to see if a falling drop with gravity and realistic properties give correct result, in opposition to dam break.

Remarks: 1- note that A12 and A21 are reassembled in smagorinsky and not in Galerkin, this is because they depend on turbulent viscosity, which varies in time, in contrary to Galerkin. In Galerkin only A11 adn A22 are reassembled, because of new u in convection term. It means, with VOF, Galerkin is not useable as such because no update of A12 and A21 with new mu field! Only smagorinsky routines can be used, with turbulent viscosity = 0 if laminar flow.

2- putting big viscosity ratio will still give correct pressure drop in example 43, but u is 1000 order of magnitudes higher, e-2 instead of e-5.

3- Consider dividing all the equations by rho, as in Yue et al 2003. This could improve numerics and computational time (better matrix conditioning?).

4- Dam break with realistic fluid properties needs necessarily Smagorinsky, becaues reynolds number too high...e.g. falling drop. In the thesis, supg is not only stabilizing equl-order elements but also high reynolds number...


Next step: because our dam break is still not comparable with literature, try out solitary wave. Example used in Yue et al 2003 + Thesis. Comparison with analytical result is possible.