enabled a 'script_mode' which only prints to the outfile, not to console
When using parmoon in scripts, the output written to console is often not
needed, because results can be found in the outfiles. Other console output
of the script itself is then lost between the many lines of parmoon output.
This can now be adjusted setting 'script_mode: true' in the input file.