we use a vector of local assembling routines now instead of just one
This will lead to greater modularization. I started with the simplest: CD2D.
The best benefits will be for NSE2D/NSE3D which should be done soon. The hope
is that the LocalAssembling classes can be reduced in size a lot and to make it
easier to add more stabilizations.