Added different control possibilities: time-dependent, space-dependent, x- and y-components.
Four booleans have been added to the default BCO database. A method to check their consistency has also been written.

The size of the control n_control and the value of rhs have been adapted accordingly. The code runs without error so far, but only on a very user-specific example.