Reworked test bco tbco.
Mysterious bug not solved: the original test was successful on my local computer but failed on the servers and on at least one other personal computer.

It was observed that after some, say n, optimization iterations, nlopt choosed different values for the control, although the solutions of the primal problem (u,p) and the cost functional are always the same until the n-th iteration (evaluation). This ends up with different results on my computer and the other systems.

Nlopt version was checked: it looks it is the same on all systems.

There is some hidden parameters affecting the computations and leading nlopt to take other values. It was also observed that the residual of the nl iterations in tnse2d are not exactly the same between my system and the other systems.

Some library is suspected to cause this weird effect, e.g., PETSc.

Until the problem is solved, the test is re-written such that it does NOT fail on the server. We just restrict the number of optimization iterations (i.e. number of primal evaluations) to n, where everything is still alright.