REMOVED GLOBAL PARAMETERS STARTTIME AND ENDTIME...
... both of them are not used anywhere anymore after this commit. Please make sure to remove them from your input files and replace them by time_start and time_end.

All the tests have been changed accordingly. The main programs compile, but they have not been tested.

IMPORTANT TO NOTE FOR TNSE2D AND TNSE3D: it is now possible to start a simulation from any time, not only 0. This is very practical when you want to restart an interrupted simulation, where your BCs or RHS depend on time. By setting time_start, you can restart correctly, without having your RHS or BCs back to t=0. (this was not possible so far, since STARTTIME and/or CURRENTTIME were forced to 0 at the beginning).