fixed issue, iterative solver for the pressure block in lsc, with MPI
in mpi mode the method TMatrix::multiply_with_transpose_from_right produces a
matrix which is stored in additive storage format. So any entry of the global
matrix is given by the sum of all its entries on the processors which know
this entry. This however can lead to zeros on the diagonal on some processors,
makeing jacobi, sor, ssor unusable as preconditioner (and solver). The fix
leads to more entries in the product matrix per processor, so that the diagonal
remains nonzero. However the resulting matrix is not stored in a consistent way
anymore, it is not possible to restore the correct global entry. A better idea
would be to do a consistency update for the matrix, but this is not
implemented. One could say, this is a hack.